feat(demoapp): move language switcher inside the centered container
Render the language dropdown per-view via renderLangSwitcher(container)
and position it absolutely at the container's top-right, instead of a
single viewport-fixed element on document.body. On wide screens it now
aligns to the 960px content column rather than the far viewport edge.
